AddPatient moved from pages module to apps module
