+++ apps/modernization-ui/src/routes/AppRoutes.tsx
@@ -8,7 +8,7 @@ import { Config } from 'config';
 import { useConfiguration } from 'configuration';
 import { Layout } from 'layout';
 import { CompareInvestigations } from 'pages/CompareInvestigations/CompareInvestigations';
-import { AddPatient } from 'pages/addPatient/AddPatient';
+import { AddPatient } from 'apps/patient/add/AddPatient';
 import { AddedPatient } from 'apps/patient/add/SuccessForm/AddedPatient';
 import { AdvancedSearch } from 'pages/advancedSearch/AdvancedSearch';
 import { Login } from 'pages/login/Login';
